Refactored contentbox model and views
`Contentbox`' previous method `get()` was replaced by Django's standard `get_or_create()`, and so the test `test_get()` was also removed.
