+++ contentbox/tests.py
@@ -2,7 +2,8 @@ from django.contrib.auth.models import Permission
 from users.models import User
 from django.test import TestCase
 
-from contentbox.models import ContentBox, DisplayContentBoxView
+from .models import ContentBox
+from .views import DisplayContentBoxView
 
 
 class ModelTestCase(TestCase):
@@ -12,16 +13,6 @@ class ModelTestCase(TestCase):
         self.assertEqual(contentbox.title, 'TEST_TITLE')
         self.assertEqual(contentbox.title, str(contentbox))
 
-    def test_get(self):
-        self.assertEqual(ContentBox.objects.count(), 0)
-        contentbox = ContentBox.get(title='TEST_TITLE')
-        self.assertEqual(ContentBox.objects.count(), 1)
-        self.assertEqual(contentbox.title, 'TEST_TITLE')
-
-        contentbox2 = ContentBox.get(title='TEST_TITLE')
-        self.assertEqual(ContentBox.objects.count(), 1)
-        self.assertEqual(contentbox, contentbox2)
-
     def test_display_view(self):
         view = DisplayContentBoxView()
         view.title = 'TEST_TITLE'
@@ -30,7 +21,7 @@ class ModelTestCase(TestCase):
         self.assertEqual(context['contentbox'].title, 'TEST_TITLE')
 
     def test_edit_without_permission(self):
-        ContentBox.get(title='TEST_TITLE')
+        ContentBox.objects.get_or_create(title='TEST_TITLE')
         response = self.client.get('/contentbox/1/edit/')
         self.assertNotEqual(response.status_code, 200)
 
@@ -41,6 +32,6 @@ class ModelTestCase(TestCase):
         permission = Permission.objects.get(codename='change_contentbox')
         user.user_permissions.add(permission)
         self.client.login(username=username, password=password)
-        ContentBox.get(title='TEST_TITLE')
+        ContentBox.objects.get_or_create(title='TEST_TITLE')
         response = self.client.get('/contentbox/1/edit/')
         self.assertEqual(response.status_code, 200)
